+++ frontend/zac-ui/libs/features/search/src/lib/features-search.component.ts
@@ -3,8 +3,6 @@ import {TableSort, Zaak} from '@gu/models';
 import {PageEvent} from '@angular/material/paginator';
 import {MapGeometry, MapMarker} from "../../../../shared/ui/components/src/lib/components/map/map";
 import {ZaakService} from "@gu/services";
-import {DOCUMENT} from "@angular/common";
-import {Router} from "@angular/router";
 
 @Component({
   selector: 'gu-features-search',
@@ -12,7 +10,7 @@ import {Router} from "@angular/router";
   styleUrls: ['./features-search.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class FeaturesSearchComponent implements OnInit{
+export class FeaturesSearchComponent {
   mapGeometries: MapGeometry[] = [];
   mapMarkers: MapMarker[] = [];
 
@@ -23,24 +21,8 @@ export class FeaturesSearchComponent implements OnInit{
 
   constructor(
     private changeDetectorRef: ChangeDetectorRef,
-    private zaakService: ZaakService,
-    private router: Router,
-    private renderer2: Renderer2,
-    @Inject(DOCUMENT) private document: Document) {
-  }
-
-  ngOnInit(): void {
-
-    /*
-     * /ui/zoeken route is now used for Oauth authorisation en can by configured via app.config.json: "redirectUri": "your_redirect_uri",
-     * 'contezza-zac-doclib' script must by appended to complete Oauth login
-     */
-    if (this.router.url.includes('session_state')) {
-      const script = this.renderer2.createElement('script');
-      script.src = '/ui/assets/contezza-zac-doclib.js';
-      this.renderer2.appendChild(this.document.body, script);
-    }
-  }
+    private zaakService: ZaakService
+  ) { }
 
   //
   // Context.
